chore(charts): move canvas-size out of hooks and trim the sizing guide
`canvas-size.ts` exports three pure functions and no hook, so `core/hooks/` was
the wrong home — it was also the only kebab-case module among the `use*` files
there. It now sits in `core/` beside `canvas-renderer.ts` and `y-axis-gutters.ts`,
and its test joins the other `core/*.test.ts` files.

Drop the device-pixel rounding note from the AGENTS.md sizing section (the reason
rounding exists is already on `backingSize`) and compress the Safari caveat to one
line. Cutting it entirely would leave the paragraph above claiming a lost context
always gets repainted, which isn't true there.

Generated-By: PostHog Code
Task-Id: 4780810c-8e0b-4d9b-9ee9-0bdb27328b52
