+++ packages/quill/packages/charts/src/core/canvas-size.ts
@@ -1,4 +1,4 @@
-import type { ChartDimensions, ChartMargins } from '../types'
+import type { ChartDimensions, ChartMargins } from './types'
 
 /** The subset of `DOMRect` the canvas sizing needs. */
 export interface SizeRect {
